update jest to work with clerk
