+++ jest.config.js
@@ -1,21 +1,14 @@
 module.exports = {
   preset: '@shelf/jest-mongodb',
-  testEnvironment: 'node',
+  // testEnvironment: 'node',
   transform: {
-    "^.+\\.tsx?$": "ts-jest"
+    '^.+\\.tsx?$': 'ts-jest'
   },
   globalSetup: '<rootDir>/mocks/setup/setup.ts',
   globalTeardown: '<rootDir>/mocks/setup/teardown.ts',
   setupFilesAfterEnv: ['<rootDir>/mocks/setup/setupAfterEnv.ts'],
-  testEnvironment: '<rootDir>/mocks/setup/environment.js',
-  moduleFileExtensions: [
-    "ts",
-    "tsx",
-    "js",
-    "jsx",
-    "json",
-    "node",
-  ],
+  testEnvironment: '<rootDir>/mocks/setup/environment.ts',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(ts|js)x?$',
   coverageDirectory: 'coverage',
   collectCoverageFrom: [
@@ -27,5 +20,5 @@ module.exports = {
     '!src/api/routes/module.*.{ts,tsx,js,jsx}',
     '!src/index.ts',
     '!src/config/*.ts'
-  ],
+  ]
 };
