+++ package.json
@@ -21,11 +21,11 @@
     "eslint-config-prettier": "~8.3.0",
     "eslint-plugin-jest": "~24.3.6",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "~27.0.6",
+    "jest": "~29.5.0",
     "nodemon": "^2.0.12",
     "prettier": "~2.3.2",
     "rimraf": "~3.0.2",
-    "ts-jest": "~27.0.3",
+    "ts-jest": "~29.1.1",
     "ts-node": "^10.4.0",
     "tslib": "^2.3.1",
     "tsutils": "~3.21.0",
